Refactor `patchDeploy`
- move `applyPatch` to `testutil/inject.go`
- refactor `patchDeploy` to `testutil.PatchDeploy()`
